First attempts at migrating to pydantic v2
Use pydantic v1 API wih v2 package

Replace `regex` with `pattern`

Use pydantic v2 API for model dumps

Go use pydantic v2 API again

Bump deps in pyproject and adjust for pydantic_settings

Attempt to migrate some pydantic v2 features

Run bump-pydantic

Update deps in requirements

Post pydantic-bump tweak number 3

Post pydantic-bump tweak number 4

Post pydantic-bump tweak number 5

Reintroduce email-validator dep via pydantic extra

Placating mypy

Use lax annotated type rather than validator

Attempt to simplify some tests for pydantic v2

Bump pydantic version in pre-commit

Bump to pydantic 2.3.0

+++ optimade/adapters/structures/utils.py
@@ -98,7 +98,7 @@ def fractional_coordinates(
         fractional[:, i] %= 1.0
         fractional[:, i] %= 1.0
 
-    return [tuple(position) for position in fractional]
+    return [tuple(position) for position in fractional]  # type: ignore
 
 
 def cell_to_cellpar(
@@ -158,12 +158,12 @@ def unit_vector(x: Vector3D) -> Vector3D:
         return None  # type: ignore[return-value]
 
     y = np.array(x, dtype="float")
-    return y / np.linalg.norm(y)
+    return y / np.linalg.norm(y)  # type: ignore
 
 
 def cellpar_to_cell(
     cellpar: list[float],
-    ab_normal: tuple[int, int, int] = (0, 0, 1),
+    ab_normal: tuple[float, float, float] = (0.0, 0.0, 1.0),
     a_direction: Optional[tuple[int, int, int]] = None,
 ) -> list[Vector3D]:
     """Return a 3x3 cell matrix from `cellpar=[a,b,c,alpha,beta,gamma]`.
@@ -219,7 +219,7 @@ def cellpar_to_cell(
     # Define rotated X,Y,Z-system, with Z along ab_normal and X along
     # the projection of a_direction onto the normal plane of Z.
     a_direction_array = np.array(a_direction)
-    Z = unit_vector(ab_normal)
+    Z = unit_vector(ab_normal)  # type: ignore
     X = unit_vector(a_direction_array - np.dot(a_direction_array, Z) * Z)
     Y = np.cross(Z, X)
